Add 'custom_payload' option to config
This addresses the 'append_info_to_payload' use case. Essentially, this
combines the `custom_payload` config and the `append_info_to_payload`
method override into a single config option.

Unfortunately, because of the way that the logging instrumentation uses
an `ActiveSupport::Notification` event to pass the payload through to
lograge, I had to perform a bit of magic to get this working. But this
might still serve a good use despite the magic.